refactor: move `ContractID` to `operations`
